jirabot/requirements.txt
langchain
langserve
langchain_openai
atlassian-python-api
slack
slackclient
slackeventsapi
python-dotenv
cachetools
